feat: Implement core functionality for event handling and rendering
- Introduced core.js to manage event binding, handling, and rendering logic.
- Added support for custom attributes (jt-actor, jt-source, jt-store) to facilitate dynamic behavior.
- Implemented HTTP request handling with error management and response processing.
- Created a debugging utility to log relevant information during development.
- Enhanced global hooks for request lifecycle management.
- Refactored handlers to improve access and registration of custom functions.
- Updated store management to validate and retrieve nested values.
- Improved template compilation and rendering logic to support dynamic content.
- Added utility functions for type checking and DOM manipulation.
