+++ docs/jtml-min.js
@@ -1,1 +1,1 @@
-var JTML=function(){"use strict";const t=(t,e)=>{if("string"!=typeof e||""===e)return t;if("."===e)return t;let n=t;for(const t of e.split(".")){if(!n)return;n=n[t]}return n};function e(...t){}function n(...t){}const r=t=>{const e=[];for(let n=0;n<t.length;n++){const r=t[n];if(r.nodeType!==Node.TEXT_NODE){if(r.nodeType===Node.ELEMENT_NODE)if(r.hasAttribute("jt-foreach"))e.push(u(r));else{if(r.hasAttribute("jt-if")){const r=o(t,n);e.push(f(r)),n=r.t-1;continue}(r.hasAttribute("jt-elseif")||r.hasAttribute("jt-else"))&&(r.hasAttribute("jt-elseif"),r.hasAttribute("jt-elseif")&&r.getAttribute("jt-elseif")),e.push(s(r))}}else e.push(c(r))}return e},o=(t,e)=>{const n=[i(t[e],1)];let r=e+1,o=!1;for(;r<t.length;){const e=t[r];if(e.nodeType===Node.ELEMENT_NODE){if(e.hasAttribute("jt-elseif")){if(o)continue;n.push(i(t[r],2))}else{if(!e.hasAttribute("jt-else"))break;if(o)continue;n.push(i(t[r],3)),o=!0}r++}else r++}return{o:n,t:r}},s=t=>{const e=t.getAttribute("jt-text");return{i:1,u:t,l:e&&d(e),p:a(t),$:r(t.childNodes)}},c=t=>({i:2,l:t.textContent}),f=t=>({i:3,o:t.o}),i=(t,e)=>{let n;switch(e){case 1:n=l(t,"jt-if");break;case 2:n=l(t,"jt-elseif");break;case 3:n=()=>!0}return{i:e,j:n,m:s(t)}},u=t=>{const e=t.getAttribute("jt-foreach");if(!e)return e;const[n,r]=e.trim().split(/\s+as\s+/);return{i:4,h:n,q:r,N:s(t)}},a=t=>{const e=[];for(let n of t.attributes)if(n.name.startsWith("jt-attr:")){const[,t]=n.name.split(":");if(!n.value)continue;const r=d(n.value);e.push((e,n)=>{e.setAttribute(t,r(n))})}return e},l=(e,n)=>{const r=e.getAttribute(n);if(!r)return()=>!1;const o=r.trim().split(/\s+/);if(1===o.length){let[t]=o;return"!"===t[0]?(t=t.slice(1),e=>!e[t]):e=>e[t]}const[s,c,f]=o,i={eq:(t,e)=>t===e,neq:(t,e)=>t!==e,gt:(t,e)=>t>e,lt:(t,e)=>t<e,gte:(t,e)=>t>=e,lte:(t,e)=>t<=e}[c];if(!i)return()=>!1;const u=e=>/^'.*'$/.test(e)?()=>e.slice(1,-1):"true"===e?()=>!0:"false"===e?()=>!1:"null"===e?()=>null:"undefined"===e?()=>{}:isNaN(Number(e))?n=>t(n,e):()=>Number(e),a=u(s),l=u(f);return t=>{const e=a(t),n=l(t);return i(e,n)}},d=n=>{const r=[];let o=0,s=0;const c=n.length,f=t=>{t>o&&r.push({i:1,T:n.slice(o,t)})};for(;s<c;){const t=n[s];if("\\"!==t||"{"!==n[s+1]&&"}"!==n[s+1]){if("{"===t){f(s);const t=s+1,c=n.indexOf("}",t);if(-1===c)return void e(JSON.stringify(n));const i=n.slice(t,c).trim();if(0===i.length)return;r.push({type:2,v:i}),s=c+1,o=s;continue}if("}"===t)return;s++}else f(s),r.push({i:1,T:n[s+1]}),s+=2,o=s}return f(c),e=>{let n="";if(1===r.length&&1===r[0].i){const o=t(e,r[0].T);return n+=o?o+"":r[0].T,n}for(const o of r)n+=1===o.i?o.T:t(e,o.v)+"";return n}},p=(e,n,r)=>{const{i:o}=e;if(2===o&&r){const{l:t}=e;return document.createTextNode(t)}if(3===o){const{o:t}=e;for(const e of t)if(e.j(n))return p(e.m,n,r);return}if(4===o){const{h:o,q:s,N:c}=e,f=t(n,o);if(!Array.isArray(f))return null;const i=document.createDocumentFragment();for(const t of f){const e=c.u.cloneNode(!1);n[s]=t;for(const t of c.$){const o=p(t,n,r);o&&e.appendChild(o)}i.appendChild(e)}return delete n[s],i}if(1===o){const{u:t,l:o,p:s,$:c}=e,f=r?t.cloneNode(!1):t;o&&(f.textContent=o(n));for(const t of s)t(f,n);for(const t of c){const e=p(t,n,r);r&&e&&f.appendChild(e)}return f}},$={},j={add:t=>{if("function"!=typeof t)return;const e=t();return"object"==typeof e?($[t.name]=e,()=>delete $[t.name]):void 0},set:(t,e)=>{((t,e,n)=>{if("string"!=typeof paths||""===paths)return t;const r=e.split(".");let o=t;for(let t=0;t<r.length-1;t++){const e=r[t];if(!o[e])return;o=o[e]}o[r[r.length-1]]=n})($,t,e)},get:e=>t($,e)},y={},m={add:t=>{if("function"!=typeof t)return;const e=new t;return y[e.constructor.name]=e,()=>delete y[e.constructor.name]},access:async(t,...e)=>{if(!t)return;if("string"!=typeof t)return;const[n,r=""]=t.split("."),o=y[n];return o&&o[r]&&"function"==typeof o[r]?o[r](...e):void 0}},h={beforeRequest:[],afterRequest:[],requestError:[]},w=(t,...e)=>{for(const n of h[t])n(...e)},b={apply:(t=document.body)=>{const e=t.querySelectorAll("[jt-actor]");for(const t of e)t._redered||(g(t),t._redered=!0)},run:t=>q(t,""),store:j,handlers:m,globalHooks:{register:t=>{if(t&&"object"==typeof t)for(const e in h)void 0!==t[e]&&("function"==typeof t[e]?h[e].push(t[e]):n(JSON.stringify(t)))}}},g=t=>{const e=t.getAttribute("jt-actor").split(",").map(t=>{let[e,n=""]=t.split(":");return[e.trim(),n.trim()]});for(const[n,r]of e){const e=N(t);"load"!==n?t.addEventListener(n,n=>q(t,r,e,n)):q(t,r,e)}},q=async(t,e,n,r)=>{if(r&&r.preventDefault(),!1===await m.access(e,t,r))return;if(["FORM","A"].includes(t.tagName)){const e=await R(t);if(!e)return;const r=t.getAttribute("jt-store");return r&&j.set(r,e),void x(t,n,e)}let o={};const s=j.get(t.getAttribute("jt-source"));s&&(o=s),x(t,n,o)},x=(t,e,n)=>{e&&e(n);const r=D(t,"jt-after",!0)??[];for(const t of r){const e=N(t);q(t,"",e)}},N=t=>{const e=T(t);if(!e)return;const n=D(t,"jt-target")||t,r=O(t);return o=>v(t,e,o,r,n)},T=t=>{if(t.hasAttribute("jt-html"))return t=>{const e=(new DOMParser).parseFromString(t,"text/html"),n=[],r=e.body.querySelectorAll("script");for(const t of r)n.push(t.innerHTML),t.remove();const o=document.createDocumentFragment();for(const t of e.body.children)o.appendChild(t);return{O:o,R:n}};const e=D(t,"jt-render");if(!e)return;if(e._compiled)return e._compiled;const n=(t=>{const e=t instanceof HTMLTemplateElement,n=r(e?t.content.childNodes:[t]);return function(t){const r=e?document.createDocumentFragment():null;for(const o of n){const n=p(o,t,e);e&&n&&r.appendChild(n)}return r}})(e);return e._compiled=n,n},v=(t,e,n,r,o)=>{const s=e(n);if(s){if(s instanceof DocumentFragment)return r(o,s),void b.apply(o);r(o,s.O);for(const t of s.R)Function(t)();b.apply(o)}},O=t=>({replace:(t,e)=>t.replaceChildren(e),append:(t,e)=>t.appendChild(e),prepend:(t,e)=>t.prepend(e)}[t.getAttribute("jt-swap")||"replace"]),R=async t=>{const{url:e,options:n}=C(t);try{w("beforeRequest",t,n),await m.access(t.getAttribute("jt-request:before"),t,n);const r=await fetch(e,n),o=await M(r);if(w("afterRequest",t,r,o),t.getAttribute("jt-request:after"),await m.access(t.getAttribute("jt-request:after"),t,r,o),!r.ok)throw{status:r.status,body:o};return o}catch(e){w("requestError",t,e),await m.access(t.getAttribute("jt-request\\:error"),t,e)}},C=t=>{let e=t.getAttribute("action")||t.getAttribute("href");const n=(t.getAttribute("method")||"GET").toUpperCase(),r={method:n,headers:{}};if("FORM"===t.tagName&&["POST","PUT","PATCH"].includes(n)){const e=F(t);r.body=JSON.stringify(e),r.headers={"Content-Type":"application/json"}}if("GET"===n&&"FORM"===t.tagName){const n=new FormData(t),r=new URLSearchParams(n),o=new URL(e,window.location.href);o.search=r,e=""+o}return{url:e,options:r}},F=t=>{const e=new FormData(t),n={};for(const[t,r]of e.entries())n[t]=r;return n},M=t=>t.headers.get("Content-Type").includes("text/html")?t.text():t.json(),D=(t,e,n=!1)=>{const r=t.getAttribute(e);if(!r)return null;try{return n?document.querySelectorAll(r):document.querySelector(r)}catch{return null}};return"loading"===document.readyState?document.addEventListener("DOMContentLoaded",()=>b.apply()):b.apply(),b}();
+var JTML=function(){"use strict";const t=(t,n)=>{if(!e(n))return;if("."===n)return t;let r=t;for(const t of n.split(".")){if(!r)return;r=r[t]}return r},e=t=>"string"==typeof t&&""!==t,n=t=>"function"==typeof t,r=t=>t instanceof DocumentFragment,o=Node.TEXT_NODE,s=Node.ELEMENT_NODE,c=t=>{const e=[];for(let n=0;n<t.length;n++){const r=t[n],c=t[n].nodeType;if(c!==o){if(c===s)if(r.hasAttribute("jt-foreach"))e.push(d(r));else{if(r.hasAttribute("jt-if")){const r=f(t,n);e.push(a(r)),n=r.t-1;continue}(r.hasAttribute("jt-elseif")||r.hasAttribute("jt-else"))&&(r.hasAttribute("jt-elseif"),r.hasAttribute("jt-elseif")&&r.getAttribute("jt-elseif")),e.push(i(r))}}else e.push(u(r))}return e},f=(t,e)=>{const n=[l(t[e],1)];let r=e+1,o=!1;for(;r<t.length;){const e=t[r];if(e.nodeType===s){if(e.hasAttribute("jt-elseif")){if(o)continue;n.push(l(t[r],2))}else{if(!e.hasAttribute("jt-else"))break;if(o)continue;n.push(l(t[r],3)),o=!0}r++}else r++}return{o:n,t:r}},i=t=>{const e=t.getAttribute("jt-text");return{i:1,u:t,l:e&&j(e),$:$(t),p:c(t.childNodes)}},u=t=>({i:2,l:t.textContent}),a=t=>({i:3,o:t.o}),l=(t,e)=>{let n;return n=1===e?p(t,"jt-if"):2===e?p(t,"jt-elseif"):()=>!0,{i:e,j:n,m:i(t)}},d=t=>{const e=t.getAttribute("jt-foreach");if(!e)return e;const[n,r]=e.trim().split(/\s+as\s+/);return{i:4,h:n,v:r,T:i(t)}},$=t=>{const e=[];for(const n of t.attributes)if(n.name.startsWith("jt-attr:")){if(!n.value)continue;const t=j(n.value),r=n.name.slice(8);e.push((e,n)=>{e.setAttribute(r,t(n))})}return e},p=(e,n)=>{const r=e.getAttribute(n);if(!r)return()=>!1;const o=r.trim().split(/\s+/);if(1===o.length){let[t]=o;return"!"===t[0]?(t=t.slice(1),e=>!e[t]):e=>e[t]}const[s,c,f]=o,i={eq:(t,e)=>t===e,neq:(t,e)=>t!==e,gt:(t,e)=>t>e,lt:(t,e)=>t<e,gte:(t,e)=>t>=e,lte:(t,e)=>t<=e}[c];if(!i)return()=>!1;const u=e=>{if(/^'.*'$/.test(e))return()=>e.slice(1,-1);if(!isNaN(Number(e)))return()=>Number(e);const n={true:!0,false:!1,undefined:void 0,null:null},r=n[e];return n.hasOwnProperty(e)?()=>r[e]:n=>t(n,e)},a=u(s),l=u(f);return t=>{const e=a(t),n=l(t);return i(e,n)}},j=e=>{const n=[];let r=0,o=0;const s=e.length,c=t=>{t>r&&n.push({i:1,q:e.slice(r,t)})};for(;o<s;){const t=e[o];if("\\"!==t||"{"!==e[o+1]&&"}"!==e[o+1]){if("{"===t){c(o);const t=o+1,s=e.indexOf("}",t);if(-1===s)return;const f=e.slice(t,s).trim();if(0===f.length)return;n.push({i:2,C:f}),o=s+1,r=o;continue}if("}"===t)return;o++}else c(o),n.push({i:1,q:e[o+1]}),o+=2,r=o}return c(s),e=>{let r="";if(1===n.length&&1===n[0].i){const o=t(e,n[0].q);return r+=o?o+"":n[0].q,r}for(const o of n)r+=1===o.i?o.q:t(e,o.C)+"";return r}},m=(e,n,r)=>{const{i:o}=e;if(2===o&&r){const{l:t}=e;return document.createTextNode(t)}if(3===o){const{o:t}=e;for(const e of t)if(e.j(n))return m(e.m,n,r);return}if(4===o){const{h:o,v:s,T:c}=e,f=t(n,o);if(!Array.isArray(f))return null;const i=document.createDocumentFragment();for(const t of f){const e=c.u.cloneNode(!1);n[s]=t;for(const t of c.p){const o=m(t,n,r);o&&e.appendChild(o)}i.appendChild(e)}return delete n[s],i}if(1===o){const{u:t,l:o,$:s,p:c}=e,f=r?t.cloneNode(!1):t;o&&(f.textContent=o(n));for(const t of s)t(f,n);for(const t of c){const e=m(t,n,r);r&&e&&f.appendChild(e)}return f}},y={},h={add:t=>{if(!n(t))return;const e=t.name;if(y[e])return;const r=new t;return y[e]=r,()=>delete y[e]},access:async(t,...r)=>{if(!e(t))return;const[o,s=""]=t.split("."),c=y[o];return c&&n(c[s])?c[s](...r):void 0}},w={},b={add:t=>{if(!n(t))return;const e=t();return"object"==typeof e?(w[t.name]=e,()=>delete w[t.name]):void 0},set:(t,n)=>{((t,n,r)=>{if(!e(n))return;const o=n.split(".");let s=t;for(let t=0;t<o.length-1;t++){const e=o[t];if(!s[e])return;s=s[e]}s[o[o.length-1]]=r})(w,t,n)},get:e=>t(w,e)},g="beforeRequest",v="afterRequest",x="requestError",T={[g]:[],[v]:[],[x]:[]},q={register:(t,e)=>{if([g,v,x].includes(t))return n(e)?void(T[t]=e):(key,void key)}},C=t=>{for(const e of t.getAttribute("jt-actor").split(",")){const n=(e+":").indexOf(":"),r=e.slice(0,n).trim(),o=e.slice(n+1).trim(),s=N(t);"load"!==r?t.addEventListener(r,e=>M(t,o,s,e)):M(t,o,s)}},M=async(t,e,n,r)=>{if(r&&r.preventDefault(),e&&!1===await h.access(e,t,r))return;if("FORM"===t.tagName||"A"===t.tagName){const e=await D(t);if(!e)return;const r=t.getAttribute("jt-store");return r&&b.set(r,e),void F(t,n,e)}let o={};const s=t.getAttribute("jt-source");if(s){const t=b.get(s);t&&(o=t)}F(t,n,o)},F=(t,e,n)=>{e&&e(n);const r=A(t,"jt-after",!0)??[];for(const t of r){const e=N(t);M(t,"",e)}},N=t=>{const e=O(t);if(!e)return;const n=A(t,"jt-target")||t,r=L(t);return o=>R(t,e,o,r,n)},O=t=>{if(t.hasAttribute("jt-html"))return t=>{const e=(new DOMParser).parseFromString(t,"text/html"),n=[],r=e.body.querySelectorAll("script");for(const t of r)n.push(t.innerHTML),t.remove();const o=document.createDocumentFragment();for(const t of e.body.children)o.appendChild(t);return{M:o,F:n}};const e=A(t,"jt-render");if(!e)return;if(e._compiled)return e._compiled;const n=(t=>{const e=r(t),n=c(e?t.content.childNodes:[t]);return function(t){const r=e?document.createDocumentFragment():null;for(const o of n){const n=m(o,t,e);e&&n&&r.appendChild(n)}return r}})(e);return e._compiled=n,n},R=(t,e,n,o,s)=>{const c=e(n);if(c){if(r(c))return o(s,c),void JTML.apply(s);o(s,c.M);for(const t of c.F)Function(t)();JTML.apply(s)}},L=t=>({replace:(t,e)=>t.replaceChildren(e),append:(t,e)=>t.appendChild(e),prepend:(t,e)=>t.prepend(e)}[t.getAttribute("jt-swap")||"replace"]),D=async t=>{const{N:e,O:n}=P(t);try{((...t)=>{for(const e of T[g])e(...t)})(t,n),await h.access(t.getAttribute("jt-request:before"),t,n);const r=await fetch(e,n),o=await J(r);if(((...t)=>{for(const e of T[v])e(...t)})(t,r,o),await h.access(t.getAttribute("jt-request:after"),t,r,o),!r.ok)throw{status:r.status,body:o};return o}catch(e){((...t)=>{for(const e of T[x])e(...t)})(t,e),await h.access(t.getAttribute("jt-request:error"),t,e)}},P=t=>{const e=(t.getAttribute("method")||"GET").toUpperCase(),n={method:e,headers:{}};if("FORM"===t.tagName&&["POST","PUT","PATCH"].includes(e)){const e=k(t);n.body=JSON.stringify(e),n.headers={"Content-Type":"application/json"}}let r=t.getAttribute("action")||t.getAttribute("href");if("GET"===e&&"FORM"===t.tagName){const e=new FormData(t),n=new URLSearchParams(e),o=new URL(r,window.location.href);o.search=n,r=""+o}return{N:r,O:n}},k=t=>{const e=new FormData(t),n={};for(const[t,r]of e.entries())n[t]=r;return n},J=t=>t.headers.get("Content-Type").includes("text/html")?t.text():t.json(),A=(t,e,n=!1)=>{const r=t.getAttribute(e);if(r)try{return n?document.querySelectorAll(r):document.querySelector(r)}catch{return}},E=(t=document)=>{performance.now();const e=t.querySelectorAll("[jt-actor]");for(const t of e)t._redered||(C(t),t._redered=!0);performance.now()};return"loading"===document.readyState?document.addEventListener("DOMContentLoaded",()=>E()):E(),{apply:E,run:t=>M(t,""),store:b,handlers:h,globalHooks:q}}();
